# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/cash/current": {
      "get": {
        "operationId": "cash_current",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashSessionEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: cash.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Caja abierta con lo esperado en el cajón. `null` si está cerrada",
        "tags": [
          "Caja"
        ]
      }
    },
    "/api/v1/cash/history": {
      "get": {
        "operationId": "cash_history",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashSessionPageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: cash.read"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Historial de cajas cerradas",
        "tags": [
          "Caja"
        ]
      }
    },
    "/api/v1/cash/open": {
      "post": {
        "operationId": "cash_open",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpenCashDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashSessionEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: cash.open"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Abre la caja del día",
        "tags": [
          "Caja"
        ]
      }
    },
    "/api/v1/cash/movements": {
      "post": {
        "operationId": "cash_movement",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CashMovementDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecordCashMovementEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: cash.movement"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Anota una entrada o salida manual de efectivo",
        "tags": [
          "Caja"
        ]
      }
    },
    "/api/v1/cash/close": {
      "post": {
        "operationId": "cash_close",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseCashDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashSessionEnvelopeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido o caducado"
          },
          "403": {
            "description": "Requiere el permiso: cash.close"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Arquea y cierra. No rechaza el descuadre: lo registra",
        "tags": [
          "Caja"
        ]
      }
    },
...
  },
  "info": {
    "title": "AppSalonBelleza API",
...
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Desarrollo local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token obtenido en POST /api/v1/auth/login"
      }
    },
    "schemas": {
...
      "PageMetaResponse": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "total": {
            "type": "number",
            "example": 137,
            "description": "Total de elementos que cumplen el filtro"
          },
          "totalPages": {
            "type": "number",
            "example": 7
          },
          "hasNext": {
            "type": "boolean",
            "example": true
          },
          "hasPrevious": {
            "type": "boolean",
            "example": false
          }
        },
        "required": [
          "page",
          "limit",
          "total",
          "totalPages",
          "hasNext",
          "hasPrevious"
        ]
      },
...
      "CashMovementResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "concept": {
            "type": "string"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "type",
          "amount",
          "currency",
          "concept",
          "reference",
          "notes",
          "occurredAt"
        ]
      },
      "CashSessionResponse": {
        "type": "object",
        "properties": {
          "movements": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashMovementResponse"
            }
          },
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "openedAt": {
            "format": "date-time",
            "type": "string"
          },
          "openedById": {
            "type": "string"
          },
          "closedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "closedById": {
            "type": "string",
            "nullable": true
          },
          "openingFloat": {
            "type": "string"
          },
          "cashSales": {
            "type": "string"
          },
          "netMovements": {
            "type": "string"
          },
          "expectedAmount": {
            "type": "string"
          },
          "countedAmount": {
            "type": "string",
            "nullable": true
          },
          "difference": {
            "type": "string",
            "nullable": true
          },
          "currency": {
            "type": "string"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "movements",
          "id",
          "status",
          "openedAt",
          "openedById",
          "closedAt",
          "closedById",
          "openingFloat",
          "cashSales",
          "netMovements",
          "expectedAmount",
          "countedAmount",
          "difference",
          "currency",
          "notes"
        ]
      },
      "CashSessionEnvelopeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CashSessionResponse"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "CashSessionPageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CashSessionResponse"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaResponse"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "OpenCashDto": {
        "type": "object",
        "properties": {
          "openingFloat": {
            "type": "number",
            "example": 500,
            "description": "Fondo con el que se abre el cajón"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "openingFloat"
        ]
      },
      "CashMovementDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "CASH_IN",
              "CASH_OUT",
              "WITHDRAWAL",
              "EXPENSE",
              "CORRECTION"
            ]
          },
          "amount": {
            "type": "number",
            "example": 25.5,
            "description": "Siempre positivo; el sentido lo da el tipo"
          },
          "concept": {
            "type": "string",
            "example": "Compra de cambio en el banco"
          },
          "reference": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "amount",
          "concept"
        ]
      },
      "RecordCashMovementResponse": {
        "type": "object",
        "properties": {
          "session": {
            "$ref": "#/components/schemas/CashSessionResponse"
          },
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "concept": {
            "type": "string"
          },
          "reference": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "occurredAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "session",
          "id",
          "type",
          "amount",
          "currency",
          "concept",
          "reference",
          "notes",
          "occurredAt"
        ]
      },
      "RecordCashMovementEnvelopeResponse": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RecordCashMovementResponse"
          }
        },
        "required": [
          "data"
        ]
      },
      "CloseCashDto": {
        "type": "object",
        "properties": {
          "countedAmount": {
            "type": "number",
            "example": 812.4,
            "description": "Efectivo contado físicamente"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "countedAmount"
        ]
      },
...
    }
  }
}